Replaced PHP5.6- warning with PHP7.2- warning
PHP 7.2 and older versions are not supported by the PHP community anymore, it is recommended to move to PHP 7.4 or PHP8. With this message, I would like to prepare the plugin to be able to start using PHP 7.3+ specific features. I also would like to help WordPress users to move to a more modern tech stack if possible.