Update woocommerce.php
The current solution for getting the product categories is inconsistent. It will always retrieve the product category that is last in a sorted alphabetical order. Let's say you have a hierarchy with 3 levels, and that you have two products. These two products have these categories selected:

1: Cars / Used / Audi
2: Cars / Used / Xpeng

The first product will be logged in Google Analytics under category name "Used" and the second under category name "Xpeng". If you set fullpath to true, this is what you will get in Analytics:

1: Cars / Used
2: Cars / Used / Xpeng

This behaviour is inconsistent and makes it hard to work on reports in Analytics. I propose that the code should always use the category that is lowest in the hierarchy when returning the name or full path, so that the first scenario will log "Audi" and "Xpeng" and the second will log the full path for the whole hierarchy in all cases.

PS: The code I changed is NOT tested in production. If you choose to implement it, please do some tests :)