Update woocommerce.php
I have looked into it more, and I believe that these two small changes will be all that is needed to handle all situations of checked off items in the hierarchy. If only one is checked, you will get that one. If more are checked, you will get all that are checked in hierarchical order. That is, you will get an array where the topmost term is the first item, the second in the hierarchy is the second and so on. With array_pop you will get the bottom term in the hierarchy.

I feel that it would be a good thing to support all cases, since there can be a lot of people working on this in a large store, and some will check all while others will not. Also, some stores might need all checked off in any case. The change I did on line 96 will return the topmost brand name when you have stores that need a hierarchical brand structure. So it will be the actual brand that get the logged revenue. I remember I have a very old support case on this that I let go stale.

What to you think about this instead of the other commit?