enhanced ecommerce - unset global state var
unset the state variable window[ "gtm4wp_select_item_" + product_data.id ] before changing location. This is needed on e.g. Safari: If we don't unset this variable, it will be cached / will stay on true if going back with the browser's back button / history.back(). This in consequence made it impossible to revisit a certain item. The commit fixes this.